Move bits of execution to shared to allow for client prediction
